🔥 chore(server.js): remove unused import The import for initWinstonLogger() is not used in the file and can be safely removed.
🐛 fix(BlockHomeTabs.js): fix tabs array creation
The tabs array was not being created correctly, causing the tabs to not display properly. The tabs array is now created correctly based on the length of the tabs prop.

🐛 fix(Tabs.js): fix image variable assignment
The image variable was not being assigned correctly, causing the image to not display properly. The image variable is now assigned correctly based on the tab prop.

🔥 chore(displayBlock.js): comment out unused block
The ACF Home Hero block is currently unused and has been commented out.

🔥 chore(formatBlockData.js): comment out unused block
The ACF Home Hero block is currently unused and has been commented out.

✨ feat(AcfHomeHero.js): add support for hero_ctas
The Ac