🔧 chore(Column.js): add Column component A new Column component was added to the project. This component is used to render a single column in a multi-column layout. It receives attributes such as style, width, and backgroundColor, which are used to style the column.
🔧 chore(Columns.js): use Column component
The Columns component was updated to use the newly added Column component. The columns.map function was commented out as it was no longer needed.

🔧 chore(BlockTestimonial.js): fix random prop
The random prop in the BlockTestimonial component was fixed to check for a string value instead of a number.

🔧 chore(getStyles.js): fix optional chaining
The optional chaining operator was added to the attributes.style object in the getStyles function.

🔧 chore(CoreButton.js): remove width attribute
The width attribute was removed from the CoreButton component.

🔧 chore(CoreColumn.js): use Column component
The Core