🔥 chore(BlockStudentOrgs): remove unused component The BlockStudentOrgs component was removed from the project, so the file was deleted.
🐛 fix(AcfFacultyCard.js): fix import path
The import path for the FeaturedImage component was incorrect, so it was fixed.

✨ feat(AcfStudentOrgs.js): add AcfStudentOrgs component and StudentOrgFragment
The AcfStudentOrgs component was added to the project to display student organizations. The StudentOrgFragment was also added to the project to query for student organizations.

🔧 chore(index.js): add AcfStudentOrgs to export
The AcfStudentOrgs component was added to the export list in the index.js file.

🔧 chore(single-program.js): add StudentOrgFragment to query
The StudentOrgFragment was added to the query in the single-program.js file to query for student organizations.