🔧 fix(Testimonial.js): set width and height for Image component 🔧 fix(AcfTestimonialBlock.jsx): change default image size to MEDIUM in getTestimonials and getTestimonial queries
The Testimonial component now sets the width and height properties for the Image component to 320 pixels, ensuring consistent dimensions for the displayed images.

In the AcfTestimonialBlock.jsx file, the default image size for the getTestimonials and getTestimonial queries has been changed to MEDIUM. This ensures that the images fetched for testimonials will be of medium size, improving performance and optimizing the loading time.