🔧 chore(Link.js): comment out unused code and update regex for local links 🐛 fix(Link.js): update regex for local links to match new domain name
The code that was commented out is no longer needed and has been commented out to improve code readability. The regular expression for matching local links has been updated to match the new domain name "wilmington-nexjs.vercel.app" and "wilmington.edu". This ensures that the correct links are identified as local links.