🔀 chore(next.config.js): remove unnecessary redirects 🔥 refactor: remove unused files and code 🔥 refactor: remove unused fonts and layout component 🔥 refactor: remove old news and program pages 🔥 refactor: remove not-found-hold component
The changes in `next.config.js` remove unnecessary redirects that were no longer needed.

The changes in the other files involve removing unused files, fonts, and components that were no longer being used in the application. This helps to clean up the codebase and improve maintainability.

🔥 refactor(layout.js): remove unused imports and commented code
🔥 refactor(not-found.tsx): remove unused imports and commented code
🔥 refactor(page.js): remove unused imports and commented code
🔥 refactor(page.js): remove unused imports and commented code
The changes were made to remove unused imports and commented code, improving the codebase's cleanliness and reducing unnecessary clutter.

🎉 feat(archive-news-and-events.js): add new template for news and events archive
✨ feat(index.js): add new template for news and events archive
The `archive-news-and-events.js` file is added as a new template for the news and events archive. This template is responsible for rendering the news and events archive page with appropriate components and data fetching logic.

In the `index.js` file, the `ArchiveNewsAndEvents` template is imported and added as a template for both the `archive-news` and `archive-event` routes. This allows the `ArchiveNewsAndEvents` template to be used for rendering the news and events archive page for both news and events separately.

These changes are made to enhance the website's functionality by providing a dedicated template for the news and events archive, improving the user experience and making it easier to maintain and update the codebase.