🐛 fix(Layout.js): add 'first' parameter to resourceMenuItems query 🐛 fix(AlgoliaResults.js): remove unused import and comment 🐛 fix(AlgoliaSearch.js): remove unused import and comment 🐛 fix(Search.js): remove unused import and parameter
In Layout.js, the 'resourceMenuItems' query now includes a 'first' parameter with a value of 20. This limits the number of menu items returned to 20.

In AlgoliaResults.js, an unused import and comment have been removed.

In AlgoliaSearch.js, an unused import and comment have been removed.

In Search.js, an unused import and parameter have been removed.