🔧 chore(next-sitemap.config.js): update siteUrl to 'https://www.wilmington.edu/' 🔧 chore(next.config.js): reorganize redirects array to include custom redirect 🔧 chore(Link.js): add regex for replacing wilmington.edu/wp-content/uploads with wordpress.wilmington.edu/wp-content/uploads
The siteUrl in next-sitemap.config.js is updated to 'https://www.wilmington.edu/' to reflect the correct URL for the website. This ensures that the sitemap generated by next-sitemap includes the correct URL.

In next.config.js, the redirects array is reorganized to include a custom redirect for '/wp-content/uploads/:slug*'. This redirect ensures that requests to this path are permanently redirected to 'https://wordpress.wilmington.edu/wp-content/uploads/:slug'.

In Link.js, a new regex is added to the regexList to replace occurrences of 'wilmington.edu/wp-content/uploads' with 'wordpress.wilmington.edu/wp-content/uploads'.