🎨 style(HomeHero.module.scss): update ::before pseudo-element styles 🎨 style(_homepage.scss): update margin-top for .intro-content
The changes in HomeHero.module.scss update the styles for the ::before pseudo-element of the HomeHero component. The width is changed to 100vw and the position is centered using left: 50% and transform: translateX(-50%). Additionally, a media query is added to reset the styles for larger screens.

The changes in _homepage.scss update the margin-top for the .intro-content class in the intro-home-page section. The margin-top is updated to use the clamp function to ensure a responsive margin based on the viewport width.