🔀 refactor(MainNavigation.js): rename useLayoutData to useMenuData 🔀 refactor(AlertBar.tsx): rename useLayoutData to useCustomData 🔀 refactor(Footer.js): rename useLayoutData to useMenuData 🔀 refactor(Header.js): rename useLayoutData to useMenuData 🔀 refactor(CustomSettingsProvider.tsx): add useCustomData hook 🔀 refactor(MenuProvider.tsx): add useMenuData hook 🔀 refactor(index.js): update imports and exports 🔀 refactor(fetchCustomSettings.js): remove unused file 🔀 refactor(allMenus.js): remove unused file
The changes were made to improve code readability and maintainability by using more descriptive and accurate function names. The use of the `useLayoutData` hook was replaced with the more appropriate `useMenuData` hook in the respective components. The `useCustomData` hook was added to

🔧 fix(_app.jsx): remove unused imports and commented out code
🚀 feat(_app.jsx): add logToConsole function call
The unused imports and commented out code have been removed to improve code cleanliness. The logToConsole function is now called to display a welcome message in the console.